requirements_yolo.txt
fastapi
uvicorn[standard]
ultralytics
Pillow
python-multipart